Code climate: backend/sdoc: dissolve type_system into grammar_element and document_view
Besides cleaner code, this simplifies tracing the L2 requirements.
